Fixed a major bug with ignoring invisible elements
I was originally using a jQuery filter (:visible) to see if we should keep
traversing the DOM. I assumed it was just checking if the display == none,
but it was actually also checking if the size was zero. This normally isn't
a problem, but some crappy sites (often generated using some crappy language,
like ASP) had divs with width and height 0 that contained content. Explicitly
checking for display == none fixed the problem